create ddpg training setup, minor refactoring
