--- src/stock_trading.py
@@ -1,154 +1,0 @@
-"""
-Use DDPG to train a stock trader based on a window of history price
-"""
-
-from __future__ import print_function, division
-
-import numpy as np
-import argparse
-import pprint
-
-from model.ddpg.actor import StockActor
-from model.ddpg.critic import StockCritic
-from model.ddpg.ddpg import DDPG
-from model.ddpg.ornstein_uhlenbeck import OrnsteinUhlenbeckActionNoise
-
-from environment.portfolio import PortfolioEnv
-from utils.data import read_stock_history, normalize
-
-import torch
-import torch.nn as nn
-
-
-
-def get_model_path(window_length, predictor_type, use_batch_norm):
-    if use_batch_norm:
-        batch_norm_str = 'batch_norm'
-    else:
-        batch_norm_str = 'no_batch_norm'
-    return 'weights/stock/{}/window_{}/{}/checkpoint.ckpt'.format(predictor_type, window_length, batch_norm_str)
-
-
-def get_result_path(window_length, predictor_type, use_batch_norm):
-    if use_batch_norm:
-        batch_norm_str = 'batch_norm'
-    else:
-        batch_norm_str = 'no_batch_norm'
-    return 'results/stock/{}/window_{}/{}/'.format(predictor_type, window_length, batch_norm_str)
-
-
-def get_variable_scope(window_length, predictor_type, use_batch_norm):
-    if use_batch_norm:
-        batch_norm_str = 'batch_norm'
-    else:
-        batch_norm_str = 'no_batch_norm'
-    return '{}_window_{}_{}'.format(predictor_type, window_length, batch_norm_str)
-
-def test_model(env, model):
-    observation, info = env.reset()
-    done = False
-    while not done:
-        action = model.predict_single(observation)
-        observation, _, done, _ = env.step(action)
-    env.render()
-
-
-def test_model_multiple(env, models):
-    observation, info = env.reset()
-    done = False
-    while not done:
-        actions = []
-        for model in models:
-            actions.append(model.predict_single(observation))
-        actions = np.array(actions)
-        observation, _, done, info = env.step(actions)
-    env.render()
-
-def peek_stock_data():
-    # Open the H5 file
-    with h5py.File('utils/datasets/stocks_history_target_2.h5', 'r') as f:
-        # Print the keys in the file
-        print("Keys in the H5 file:", list(f.keys()))
-        
-        # Get the stock data and abbreviations
-        stock_data = f['history'][:]  # Assuming 'stock_data' is the dataset name
-        stock_abbr = f['abbreviation'][:] # Assuming 'abbreviation' is the dataset name
-        
-        # Convert byte strings to regular strings if needed
-        stock_abbr = [abbr.decode('utf-8') if isinstance(abbr, bytes) else abbr for abbr in stock_abbr]
-        
-        # Print the shape of the data
-        print("\nData shape:", stock_data.shape)
-        print("Stock abbreviations:", stock_abbr)
-        
-        # Show first few entries
-        print("\nFirst few entries of the stock data:")
-        # Assuming the data structure is [stocks, time_steps, features]
-        for i in range(min(5, len(stock_abbr))):
-            print(f"\n{stock_abbr[i]}:")
-            print(stock_data[i, :5])  # First 5 time steps
-
-if __name__ == '__main__':
-
-    parser = argparse.ArgumentParser(description='Provide arguments for training different DDPG models')
-
-    parser.add_argument('--debug', '-d', help='print debug statement', default=False)
-    parser.add_argument('--predictor_type', '-p', help='cnn or lstm predictor', required=True)
-    parser.add_argument('--window_length', '-w', help='observation window length', required=True)
-    parser.add_argument('--batch_norm', '-b', help='whether to use batch normalization', required=True)
-
-    args = vars(parser.parse_args())
-
-    pprint.pprint(args)
-
-    if args['debug'] == 'True':
-        DEBUG = True
-    else:
-        DEBUG = False
-
-    history, abbreviation = read_stock_history(filepath='utils/datasets/stocks_history_target.h5')
-    history = history[:, :, :4]
-    target_stocks = abbreviation
-    num_training_time = 1095
-    window_length = int(args['window_length'])
-    nb_classes = len(target_stocks) + 1
-
-    # get target history
-    target_history = np.empty(shape=(len(target_stocks), num_training_time, history.shape[2]))
-    for i, stock in enumerate(target_stocks):
-        target_history[i] = history[abbreviation.index(stock), :num_training_time, :]
-
-    print(target_history.shape)
-    # setup environment
-    env = PortfolioEnv(target_history, target_stocks, steps=1000, window_length=window_length)
-
-    action_dim = [nb_classes]
-    state_dim = [nb_classes, window_length]
-    batch_size = 64
-    action_bound = 1.
-    tau = 1e-3
-    assert args['predictor_type'] in ['cnn', 'lstm'], 'Predictor must be either cnn or lstm'
-    predictor_type = args['predictor_type']
-    if args['batch_norm'] == 'True':
-        use_batch_norm = True
-    elif args['batch_norm'] == 'False':
-        use_batch_norm = False
-    else:
-        raise ValueError('Unknown batch norm argument')
-    actor_noise = OrnsteinUhlenbeckActionNoise(mu=np.zeros(action_dim))
-    model_save_path = get_model_path(window_length, predictor_type, use_batch_norm)
-    summary_path = get_result_path(window_length, predictor_type, use_batch_norm)
-
-    variable_scope = get_variable_scope(window_length, predictor_type, use_batch_norm)
-
-    actor = StockActor(state_dim, action_dim, action_bound, 1e-4, tau, batch_size,
-                        predictor_type, use_batch_norm)
-    critic = StockCritic(state_dim=state_dim, action_dim=action_dim, learning_rate=1e-3, tau=1e-3,
-                             num_actor_vars=actor.get_num_trainable_vars(),
-                             predictor_type=predictor_type, use_batch_norm=use_batch_norm)
-    ddpg_model = DDPG(env, sess, actor, critic, actor_noise,
-                        config_file='config/stock.json', model_save_path=model_save_path,
-                        summary_path=summary_path)
-    ddpg_model.initialize(load_weights=False)
-    ddpg_model.train()
-
